Writing the image
Writing an image from the original data.
Using the original image produced. The algorithm re-writes the data and
produces the new image with the same header information.
There is no more need to use resampling while re-writing the image.

The resampling needed to display the data in this viewer is LIP (left
inferior posterior)